Sum of Squares

+++ math/calculus/9-sum_total.py
@@ -6,5 +6,5 @@ def summation_i_squared(n):
     """a function that calculates the sum of squared numbers"""
     if type(n) != int:
         return None
-    sum = (n*(n+1)*(2*n+1))/6
+    sum = int((n*(n+1)*(2*n+1))/6)
     return sum
